requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
anthropic==0.7.8
python-multipart==0.0.6
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
boto3==1.34.0
twilio==8.10.0
streamlit==1.28.1
pandas==2.1.4
plotly==5.17.0
requests==2.31.0
pillow==10.1.0

# Security dependencies
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0

